feat:【数据统计】用户态数据统计- 总览 #6587

+++ src/backend/ci/core/process/api-process/src/main/kotlin/com/tencent/devops/process/pojo/PipelineFailInfoDO.kt
@@ -37,5 +37,5 @@ data class PipelineFailInfoDO(
     @ApiModelProperty("错误类型标识")
     val errorType: Int,
     @ApiModelProperty("错误次数")
-    val errorCount: Int
+    val errorCount: Long
 )
